Add files via upload
I used the command interFE for the regressions. After many experiments with different versions, I could see that the number of the fixed effects (r) does not change the results much and gsynth with its permutations is not needed. We can qualitatively look at r=0, which is only additive fixed effects, r=1, r=10, r=50, r=100. The diagnostics are included in the end.